Fixed bug for submatrix
The operator[] of submatrix returns a sequence. This sequence consists
of a matrix row and a beginning and end of that row for the sequence.
Since the beginning and end are on the row, we can give those with the m
instead of the n function.
The matrix row consists of the matrix and the row index of that matrix
instead of a submatrix.